// src/transactions/enums/transaction-status.enum.ts
export enum TransactionStatus {
    WAITING_FOR_PAYMENT = 'WAITING_FOR_PAYMENT', // [cite: 248]
    PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',       // [cite: 249]
    SHIPPING = 'SHIPPING',                       // [cite: 250]
    COMPLETED = 'COMPLETED',                     // [cite: 252]
    DISPUTED = 'DISPUTED',                       // [cite: 253]
    CANCELLED = 'CANCELLED',                     // [cite: 254]
  }